backend/requirements-core.txt
# Core Backend Dependencies for MVP
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0

# Geospatial support (optional for now)
# geoalchemy2==0.14.2
# shapely==2.0.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
reportlab==4.2.5
